Update bot.py
Added asyncio sleeps of zero into the auto-update function, in an effort to keep it from timing out discord.py, as it appears to be doing in the log files. 

This will be monitored in the future to ensure that it works.

Also, added code to stop the bot from sending messages saying that it restarted, when all that happened was a new instance of on_ready() was called.